// test/integration/test/ports/prisma/functional/methods-upsert-simple/_fixture/generated/contract.d.ts
export type StorageHash =
  StorageHashBase<'6b6e1d35d1483d42ba5d6e591c007268b198850a99795e4facc899b28e49bd46'>;
export type ExecutionHash =
  ExecutionHashBase<'b3a0c7629738fce7d43a03a126ffc34990efca75f034d52cdd67a1adfada5ca4'>;
export type ProfileHash =
  ProfileHashBase<'3916f444a8a17ad749191acf9e08dad97d1a327b88c2f1d45d12f240296aa8b2'>;

export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly User: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly user: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly user: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;

type ContractBase = Omit<
  ContractType<{
    readonly namespaces: {
      readonly public: {
        readonly id: 'public';
        readonly kind: 'postgres-schema';
        readonly entries: {
          readonly table: {
            readonly user: {
              columns: {
                readonly id: {
                  readonly nativeType: 'text';
                  readonly codecId: 'pg/text@1';
                  readonly nullable: false;
                };
                readonly name: {
                  readonly nativeType: 'text';
                  readonly codecId: 'pg/text@1';
                  readonly nullable: false;
                };
              };
              primaryKey: { readonly columns: readonly ['id'] };
              uniques: readonly [{ readonly columns: readonly ['name'] }];
              indexes: readonly [];
              foreignKeys: readonly [];
            };
          };
        };
      };
    };
    readonly storageHash: StorageHash;
  }>,
  'roots' | 'domain'
> & {
  readonly target: 'postgres';
  readonly targetFamily: 'sql';
  readonly roots: {
    readonly user: { readonly namespace: 'public' & NamespaceId; readonly model: 'User' };
  };
  readonly domain: {
    readonly namespaces: {
      readonly public: {
        readonly models: {
          readonly User: {
            readonly fields: {
              readonly id: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/text@1' };
              };
              readonly name: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/text@1' };
              };
            };
            readonly relations: Record<string, never>;
            readonly storage: {
              readonly table: 'user';
              readonly namespaceId: 'public';
              readonly fields: {
                readonly id: { readonly column: 'id' };
                readonly name: { readonly column: 'name' };
              };
            };
          };
        };
      };
    };
  };
  readonly capabilities: {
    readonly postgres: {
      readonly distinctOn: true;
      readonly jsonAgg: true;
      readonly lateral: true;
      readonly limit: true;
      readonly orderBy: true;
      readonly returning: true;
    };
    readonly sql: {
      readonly defaultInInsert: true;
      readonly enums: true;
      readonly lateral: true;
      readonly returning: true;
      readonly scalarList: true;
    };
  };
  readonly extensions: {};
  readonly execution: {
    readonly executionHash: ExecutionHash;
    readonly mutations: {
      readonly defaults: readonly [
        {
          readonly ref: {
            readonly namespace: 'public';
            readonly table: 'user';
            readonly column: 'id';
          };
          readonly onCreate: { readonly kind: 'generator'; readonly id: 'cuid2' };
        },
      ];
    };
  };
  readonly meta: {};

  readonly profileHash: ProfileHash;
};

export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;